Don't try to relaunch a dead kernel with the same ports
Restarting with the same ports can never be 100% reliable, because one
or more ports may have been taken by another process in the meantime.
So applications that want reliable automatic restart need to be prepared to deal
with the ports changing. In which case, we can reduce complexity by
starting a fresh kernel with randomly generated ports on each restart.