Let KernelManager hold kernel_id
Kernel providers will want to use the same kernel_id used by the
MappingKernelManager as a means of managing a kernel's lifecycle.
This PR adds a `kernel_id` member to the KernelManager abstract
base class and initializes its value in the KernelManager constructor.
This allows kernel providers to determine the best location for where
to set `kernel_id` so that it can be used to manage the kerne's
lifecycle.

The MappingKernelManager start logic will then derive the kernel_id
from the provider's KernelManager instance.

Fixes #10